more work towards orbit trapping
In this version, i've attempted to shade the surface by taking an orbit trap for a point that's offset from the hit surface point along the normal at that point. For the orbit trap itself, it accumulates the bisector of the difference vectors for each iteration of the function, and then uses that accumulated bisector in a dot product with teh surface normal to generate an angle, which is finally used to index a color value of the surface.